- Added KeyPress*Auto functions to auto detect extended press keys. - Separated extended key press in KeyPressCombo functions.

+++ Desktop/Functions/AVInterface.cs
@@ -41,7 +41,7 @@ namespace ArnoldVinkCode
                             if (Keyboard.FocusedElement == null)
                             {
                                 Debug.WriteLine("Failed focusing on the element after " + whileLoopCount + " times, pressing tab key.");
-                                await KeySendSingle((byte)KeysVirtual.Tab, mainWindowHandle);
+                                await KeySendSingle(KeysVirtual.Tab, mainWindowHandle);
                             }
                             return;
                         }
@@ -55,13 +55,13 @@ namespace ArnoldVinkCode
                 else
                 {
                     Debug.WriteLine("Focus element cannot be focused on, pressing tab key.");
-                    await KeySendSingle((byte)KeysVirtual.Tab, mainWindowHandle);
+                    await KeySendSingle(KeysVirtual.Tab, mainWindowHandle);
                 }
             }
             catch
             {
                 Debug.WriteLine("Failed focusing on the element, pressing tab key.");
-                await KeySendSingle((byte)KeysVirtual.Tab, mainWindowHandle);
+                await KeySendSingle(KeysVirtual.Tab, mainWindowHandle);
             }
         }
     }
